% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaNB-package.R
\docType{package}
\name{MetaNB-package}
\alias{MetaNB}
\alias{MetaNB-package}
\title{MetaNB: Meta-Analysis of Net Benefit and Value-of-Information}
\description{
Performs Bayesian trivariate random-effects meta-analysis of prevalence, sensitivity, and specificity to estimate net benefit across heterogeneous validation settings (Wynants et al., 2018) \doi{10.1002/sim.7653}. Using Markov chain Monte Carlo sampling via JAGS, it produces estimates of net benefit and its credible and prediction intervals, and the posterior probability a model is clinically useful in a new setting, for given relevant harm-to-benefit ratios. It further estimates value-of-information metrics, including expected value of perfect and partial perfect information, to quantify uncertainty for the decision to implement the model in care across heterogeneous settings.
}
\author{
\strong{Maintainer}: Junfeng Wang \email{wangjunfeng7@gmail.com} (\href{https://orcid.org/0000-0001-5157-5355}{ORCID})

Authors:
\itemize{
  \item Kim (Zhipei) Wang \email{wzpskye@gmail.com} (\href{https://orcid.org/0009-0002-9839-9138}{ORCID})
  \item Laure Wynants \email{laure.wynants@maastrichtuniversity.nl} (\href{https://orcid.org/0000-0002-3037-122X}{ORCID})
}

}
\keyword{internal}
